Changed FightTerrain x y  from int -> short

+++ Progemon/src/test/TestFileUtitlity.java
@@ -43,7 +43,7 @@ public class TestFileUtitlity {
 		}
 
 		// Check HashMap properties
-		
+
 		Pokedex.printPokedex();
 
 		assertEquals("Length = 28", 28, Pokedex.getPokedex().size());
@@ -57,7 +57,7 @@ public class TestFileUtitlity {
 		assertTrue("Contains Charmander", Pokedex.getPokedex().containsValue("Charmander"));
 		assertTrue("Contains Charmander", Pokedex.getPokedex().containsKey(4));
 		assertEquals("ID 004 = Charmander", Pokedex.getPokedex().get(4), "Charmander");
-		
+
 		assertTrue("Contains Pikachu", Pokedex.getPokedex().containsValue("Pikachu"));
 		assertTrue("Contains Pikachu", Pokedex.getPokedex().containsKey(25));
 		assertEquals("ID 025 = Pikachu", Pokedex.getPokedex().get(25), "Pikachu");
@@ -83,19 +83,20 @@ public class TestFileUtitlity {
 		assertEquals("Height = 6", 6, fightMap.length);
 
 		// GRASS GRASS GRASS GRASS GRASS GRASS GRASS GRASS
-		assertTrue("Top Left Terrain", new FightTerrain(0, 0, FightTerrain.TerrainType.GRASS).equals(fightMap[0][0]));
+		assertTrue("Top Left Terrain",
+				new FightTerrain((short) 0, (short) 0, FightTerrain.TerrainType.GRASS).equals(fightMap[0][0]));
 
-		for (int i = 1; i <= 7; i++) {
+		for (short i = 1; i <= 7; i++) {
 			assertTrue("First Row, Column : " + i,
-					new FightTerrain(i, 0, FightTerrain.TerrainType.GRASS).equals(fightMap[0][i]));
+					new FightTerrain(i, (short) 0, FightTerrain.TerrainType.GRASS).equals(fightMap[0][i]));
 		}
 
 		// GRASS ROCK ROCK TREE WATER WATER GRASS GRASS
 		assertTrue("Second Row First Column",
-				new FightTerrain(0, 1, FightTerrain.TerrainType.GRASS).equals(fightMap[1][0]));
-		assertTrue(new FightTerrain(1,1, FightTerrain.TerrainType.ROCK).equals(fightMap[1][1]));
-		assertTrue(new FightTerrain(3,1, FightTerrain.TerrainType.TREE).equals(fightMap[1][3]));
-		assertTrue(new FightTerrain(5,1, FightTerrain.TerrainType.WATER).equals(fightMap[1][5]));
+				new FightTerrain((short) 0, (short) 1, FightTerrain.TerrainType.GRASS).equals(fightMap[1][0]));
+		assertTrue(new FightTerrain((short) 1, (short) 1, FightTerrain.TerrainType.ROCK).equals(fightMap[1][1]));
+		assertTrue(new FightTerrain((short) 3, (short) 1, FightTerrain.TerrainType.TREE).equals(fightMap[1][3]));
+		assertTrue(new FightTerrain((short) 5, (short) 1, FightTerrain.TerrainType.WATER).equals(fightMap[1][5]));
 
 	}
 
@@ -114,20 +115,25 @@ public class TestFileUtitlity {
 		String[] args = "001 49 49 45 45 3 1 WALK".split(" ");
 		Pokemon testFirstPokemon = new Pokemon(args);
 		Pokemon firstPokemon = Pokedex.getAllPokemons().get(0);
-		
-		assertTrue("First Pokemon Stats", Double.compare(testFirstPokemon.getBase().attackStat, firstPokemon.getBase().attackStat) == 0);
-		assertTrue("First Pokemon Stats", Double.compare(testFirstPokemon.getBase().defenceStat, firstPokemon.getBase().defenceStat) == 0);
+
+		assertTrue("First Pokemon Stats",
+				Double.compare(testFirstPokemon.getBase().attackStat, firstPokemon.getBase().attackStat) == 0);
+		assertTrue("First Pokemon Stats",
+				Double.compare(testFirstPokemon.getBase().defenceStat, firstPokemon.getBase().defenceStat) == 0);
 		assertTrue("First Pokemon Stats",
 				Double.compare(testFirstPokemon.getMoveRange(), firstPokemon.getMoveRange()) == 0);
-		assertTrue("First Pokemon Stats", Double.compare(testFirstPokemon.getBase().speed, firstPokemon.getBase().speed) == 0);
-		assertTrue("First Pokemon Stats", Double.compare(testFirstPokemon.getBase().fullHP, firstPokemon.getBase().fullHP) == 0);
+		assertTrue("First Pokemon Stats",
+				Double.compare(testFirstPokemon.getBase().speed, firstPokemon.getBase().speed) == 0);
+		assertTrue("First Pokemon Stats",
+				Double.compare(testFirstPokemon.getBase().fullHP, firstPokemon.getBase().fullHP) == 0);
 		assertEquals("First Pokemon Moves", 2, firstPokemon.getActiveSkills().size());
-		
-		
+
 		// Ivysaur Moves = 2
 		assertEquals("Ivysaur ActiveSkill", 2, Pokedex.getAllPokemons().get(1).getActiveSkills().size());
-		assertEquals("Ivysaur ActiveSkill", "Tackle", Pokedex.getAllPokemons().get(1).getActiveSkills().get(0).getName());
-		assertEquals("Ivysaur ActiveSkill", "Razor Leaf", Pokedex.getAllPokemons().get(1).getActiveSkills().get(1).getName());
+		assertEquals("Ivysaur ActiveSkill", "Tackle",
+				Pokedex.getAllPokemons().get(1).getActiveSkills().get(0).getName());
+		assertEquals("Ivysaur ActiveSkill", "Razor Leaf",
+				Pokedex.getAllPokemons().get(1).getActiveSkills().get(1).getName());
 	}
 
 	@Test
@@ -146,10 +152,9 @@ public class TestFileUtitlity {
 		assertEquals("Second Move : Razor Wind", "Razor Wind", aSkills.get(1).getName());
 		assertTrue("Second Move : Razor Wind", Double.compare(80, aSkills.get(1).getPower()) == 0);
 	}
-	
-	
+
 	@Test
-	public void testLoadActiveSkills2(){
+	public void testLoadActiveSkills2() {
 		ActiveSkill.clearAllActiveSkills();
 		String customFilePath = "test/test1loadActiveSkills.txt";
 		try {
@@ -157,11 +162,11 @@ public class TestFileUtitlity {
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
-		
+
 		ArrayList<ActiveSkill> aSkills = new ArrayList<ActiveSkill>(ActiveSkill.getAllActiveSkills());
-		
+
 		assertEquals("Size of allActiveSkills", 2, aSkills.size());
-		
+
 		assertEquals("First Move : Hello WORLD", "Hello World", aSkills.get(0).getName());
 		assertTrue("First Move : Hello WORLD", Double.compare(13, aSkills.get(0).getPower()) == 0);
 		assertEquals("Second Move : Utit Mo Ko", "Utit Mo Ko", aSkills.get(1).getName());
